Run the backup if this node is lexographically the lowest secondary.
This allows the service to be run safely on any MongoDB node in a
cluster, and it will come to the same conclusion about which node should
be used to run the backup as another instance of the service running on
any other MongoDB node in the same cluster.

Therefore, with this change we can safely run _N_ instances of the
service on all _N_ MongoDB nodes and have confidence that each time the
service is triggered, one secondary node will always be chosen
successfully for a backup. This avoids the risk that a singleton backup
service runs by chance for a long time on the MongoDB primary and avoids
taking a backup.