Use faster gzip settings
Change from default to BestSpeed setting, to reduce CPU usage during
backup and also make it run quicker.

In a quick before/after test this change made the backup run 15%
quicker, take less CPU (unscientifically, instead of 80-110% it was
hovering around 50-80%).  The tradeoff here is that the backed up size
in s3 was ~3% bigger.